+++ src/utils/time.ts
@@ -1,3 +1,7 @@
+import { CountBlocksBetweenProps as Period } from '../repositories/BlockRepository';
+import { subDays, isValid } from 'date-fns';
+
+const MAX_PERIOD_RANGE = 7 * 24 * 60 * 60;
 export const MINUTE = 'm';
 export const HOUR = 'h';
 export const DAY = 'd';
@@ -23,4 +27,31 @@ export const getDateAtMidnight = (day: string): Date => {
   date.setUTCSeconds(0);
 
   return date;
+};
+
+export const getDateOrDefault = ({ startDate, endDate }: Period<string>): Period<Date> => {
+  const end = isValid(new Date(endDate)) ? new Date(endDate) : new Date();
+  const start = isValid(new Date(startDate)) ? new Date(startDate) : subDays(end, 1);
+
+  return validatePeriod({ startDate: start, endDate: end });
+};
+
+const validatePeriod = (period: Period<Date>): Period<Date> => {
+  if (isPeriodRangeValid(period) && isPeriodBelowThreshold(period, MAX_PERIOD_RANGE)) {
+    return period;
+  }
+
+  return {
+    startDate: subDays(period.endDate, MAX_PERIOD_RANGE),
+    endDate: period.endDate,
+  };
+};
+
+const isPeriodRangeValid = (period: Period<Date>): boolean => {
+  return period.endDate.getTime() > period.startDate.getTime();
+};
+
+const isPeriodBelowThreshold = ({ startDate, endDate }: Period<Date>, threshold = MAX_PERIOD_RANGE): boolean => {
+  const currentPeriodTime = endDate.getTime() / 1000 - startDate.getTime() / 1000;
+  return currentPeriodTime < threshold;
 };